make structure changes
