--- .env.template
@@ -1,6 +1,0 @@
-# Flask server configuration.
-FLASK_APP=todo_app/app
-FLASK_ENV=development
-
-# Change the following values for local development.
-SECRET_KEY=secret-key
